Sponge constructed
Reinforced concrete now inserted into a sponge
construction, allow arbitrary lenght input and
outputs. This can be done in circuit or out of
circuit. Tests for this added, and reinforced
concrete altered to optimise for sponge construction
(inserting variables that are used repeatedly,
rather than setting them as variables in every
round)